+++ frontend/src/pages/QuizPage.tsx
@@ -1,15 +1,11 @@
 import { useEffect, useRef, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import axios from "../api/axios";
+import QuizInput from "../components/QuizInput";
 import {
-  question,
-  questionInput,
-  questionNumbers,
-  quizAnswerDiv,
   quizButton,
   quizButtonDiv,
   quizContainer,
-  quizInput,
   quizTitleContainer,
   todayBg,
   todayQuiz,
@@ -99,13 +95,7 @@ function QuizPage() {
     FetchPostQuiz();
     navigate("/result", { state: feedback });
   };
-  const handleAnswerChange = (index: number, answer: string) => {
-    // const updatedAnswers = [...answers];
-    const updatedAnswers = [...answers.current];
-    updatedAnswers[index] = answer;
-    // setAnswers(updatedAnswers);
-    answers.current = updatedAnswers;
-  };
+
   {
     return (
       <div className={quizContainer}>
@@ -120,27 +110,12 @@ function QuizPage() {
           <p className={todayQuiz}>TODAY QUIZ</p>
         </div>
 
-        {quizs.map(
-          (quiz, id) =>
-            // 현재 퀴즈 인덱스와 매핑되는 퀴즈를 보여줌
-            id === currentQuizIndex && (
-              <div key={quiz.id} style={{ width: "100%" }}>
-                <p className={questionNumbers}>{currentQuizIndex + 1}/5</p>
-                <p className={question}>{quiz.category}</p>
-                <div className={quizAnswerDiv}>
-                  <div className={questionInput}>{quiz.question}</div>
-                  <input
-                    className={quizInput}
-                    type="text"
-                    placeholder="정답을 입력하세요"
-                    onChange={(e) =>
-                      handleAnswerChange(currentQuizIndex, e.target.value)
-                    }
-                  />
-                </div>
-              </div>
-            )
-        )}
+        <QuizInput
+          quizs={quizs}
+          currentQuizIndex={currentQuizIndex}
+          answers={answers}
+        />
+        
         <div className={quizButtonDiv}>
           <button
             className={quizButton}
